Assistant/requirements.txt
annotated-types==0.7.0
appopener==1.7
audioop-lts==0.2.2
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
comtypes==1.4.13
docopt==0.6.2
Flask==3.1.2
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
gTTS==2.5.4
httplib2==0.31.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
MouseInfo==0.1.3
pillow==12.0.0
pipreqs==0.4.13
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.6.1
PyGetWindow==0.0.9
PyMsgBox==2.0.1
pyparsing==3.2.5
pyperclip==1.11.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
python-dotenv==1.2.1
pyttsx3==2.99
pytweening==1.2.0
pywhatkit==5.4
pywin32==311
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
screen_brightness_control==0.24.3
soupsieve==2.8
SpeechRecognition==3.14.4
standard-aifc==3.13.0
standard-chunk==3.13.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.4
wheel==0.45.1
wikipedia==1.4.0
WMI==1.5.1
yarg==0.1.10
